(PC-29655)[PRO] refactor: remove unnecessary code around booking adapters
