+++ pro/src/core/Bookings/types.ts
@@ -1,8 +1,4 @@
-import {
-  BookingRecapResponseModel,
-  CollectiveBookingResponseModel,
-  BookingStatusFilter,
-} from 'apiClient/v1'
+import { BookingStatusFilter } from 'apiClient/v1'
 
 export type PreFiltersParams = {
   offerVenueId: string
@@ -24,30 +20,6 @@ export type APIFilters = {
   offerId?: string
   page: number
 }
-
-export type GetFilteredBookingsRecapAdapterPayload = {
-  bookings: BookingRecapResponseModel[]
-  pages: number
-  currentPage: number
-}
-
-export type GetFilteredCollectiveBookingsRecapAdapterPayload = {
-  bookings: CollectiveBookingResponseModel[]
-  pages: number
-  currentPage: number
-}
-
-export type GetFilteredBookingsRecapAdapter = Adapter<
-  PreFiltersParams & { page?: number },
-  GetFilteredBookingsRecapAdapterPayload,
-  GetFilteredBookingsRecapAdapterPayload
->
-
-export type GetFilteredCollectiveBookingsRecapAdapter = Adapter<
-  PreFiltersParams & { page?: number },
-  GetFilteredCollectiveBookingsRecapAdapterPayload,
-  GetFilteredCollectiveBookingsRecapAdapterPayload
->
 
 export type GetUserHasBookingsAdapter = Adapter<void, boolean, boolean>
