+++ pro/src/screens/IndividualOffer/BookingsSummary/BookingsSummary.tsx
@@ -14,7 +14,7 @@ import {
 } from 'core/Bookings/constants'
 import useActiveFeature from 'hooks/useActiveFeature'
 import strokeBookingHold from 'icons/stroke-booking-hold.svg'
-import { getFilteredBookingsRecapAdapter } from 'pages/Bookings/adapters/getFilteredBookingsRecapAdapter'
+import { getFilteredIndividualBookingsAdapter } from 'pages/Bookings/adapters/getFilteredIndividualBookingsAdapter'
 import { IndividualBookingsTable } from 'screens/Bookings/BookingsRecapTable/BookingsTable/IndividualBookingsTable'
 import { DEFAULT_OMNISEARCH_CRITERIA } from 'screens/Bookings/BookingsRecapTable/Filters/constants'
 import filterBookingsRecap from 'screens/Bookings/BookingsRecapTable/utils/filterBookingsRecap'
@@ -55,16 +55,14 @@ export const BookingsSummaryScreen = ({
 
   useEffect(() => {
     const loadBookings = async () => {
-      const response = await getFilteredBookingsRecapAdapter({
+      const response = await getFilteredIndividualBookingsAdapter({
         ...DEFAULT_PRE_FILTERS,
         offerId: String(offer.id),
         bookingBeginningDate: '2015-01-01',
         bookingEndingDate: format(new Date(), FORMAT_ISO_DATE_ONLY),
       })
 
-      if (response.isOk) {
-        setBookings(response.payload.bookings)
-      }
+      setBookings(response.bookings)
     }
     // eslint-disable-next-line @typescript-eslint/no-floating-promises
     loadBookings()
